Improve handling of multi-lingual address entries
At least in Biel/Bienne there are two entries per EAGID (GWR address id)
this change takes that in to account and tries harder to match up
addresses as far as possible.

Note that if in such situations only the name field is present in OSM
and the language specific fields are missing, the address will not be
found.