Handle multiple occurrences of the same number/street
This adds support for multiple occurrences of the same number street
combination, which can happen because of different postcodes in the same
municipality on the GWR side, and additionally for OSM when the
addresses is duplicate on shops etc.

The change allows us to provide better matching and warnings for every
OSM object with an address.